Split tests into separate files (to avoid clashing singletons)
